issue #54 refactor: sign up directory separate
