+++ src/components/SignUp/Email/PasswordValidator.tsx
@@ -3,7 +3,7 @@ import PasswordCheckIcon from "@/assets/svg/ic-password-check.svg?react";
 import { Flex, Text } from "@/components/common";
 import { passwordCheckData } from "@/constants/auth";
 
-import { getPasswordTextStyle } from "@/components/SignUp/SignUpEmail.style";
+import { getPasswordTextStyle } from "@/components/SignUp/Email/SignUpEmail.style";
 
 interface validatorType {
 	password: string;
