+++ src/components/SignUp/Pet/SignUpPet.style.ts
