+++ src/components/SignUp/Profile/SignUpProfile.style.ts
@@ -38,14 +38,3 @@ export const addressInputStyle = css({
 		color: Theme.color.disabled_text,
 	},
 });
-
-export const getNextButtonStyle = (text: string) => {
-	return css({
-		all: "unset",
-		padding: "8px 14px",
-		borderRadius: "4px",
-		backgroundColor: text === "이전" ? Theme.color.disabled_text : Theme.color.brand_primary,
-		color: Theme.color.white,
-		fontWeight: 600,
-	});
-};
