+++ src/components/SignUp/SignUp.shared.style.ts
@@ -49,3 +49,14 @@ export const getTextareaStyle = (height: string) => {
 		},
 	});
 };
+
+export const getNextButtonStyle = (text: string) => {
+	return css({
+		all: "unset",
+		padding: "8px 14px",
+		borderRadius: "4px",
+		backgroundColor: text === "이전" ? Theme.color.disabled_text : Theme.color.brand_primary,
+		color: Theme.color.white,
+		fontWeight: 600,
+	});
+};
